+++ envsetup.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+st_build() {
+	go build -o sndtool ./cmd/
+}